requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp_socks==0.10.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
matrix-nio==0.25.2
matrix_client
multidict==6.4.3
packaging==25.0
pluggy==1.5.0
propcache==0.3.1
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
python-socks==2.7.1
referencing==0.36.2
rpds-py==0.24.0
sniffio==1.3.1
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
unpaddedbase64==2.1.0
uvicorn==0.34.2
yarl==1.20.0
celery==5.3.6
redis==5.0.4
pydantic-settings==2.3.4
python-jose[cryptography]>=3.4.0
python-multipart>=0.0.18
prometheus-client==0.20.0
